+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [Unreleased]
 
-- Implemented repeatable key behaviour
+- Implemented repeatable keys (including backspace)
 - Implemented currently pressed key behaviour and appearance
 - Implemented enter key behaviour
 - Removed key attribute `displayIcon`