+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## [Unreleased]
 
 - Implemented long press (including long press space for keyboard switcher)
-- Implemented repeatable keys (including backspace)
+- Implemented key repeat (including backspace)
 - Implemented currently pressed key behaviour and appearance
 - Implemented enter key behaviour
 - Removed key attribute `displayIcon`